# static/openapi.yaml
openapi: 3.0.3
info:
  title: SodhiCable MES v4.0 API
  description: |
    REST API for the SodhiCable Manufacturing Execution System.

    Aligned with **MESA-11** (Manufacturing Enterprise Solutions Association) functional
    model and the **ISA-95** (IEC 62264) enterprise-control integration standard.

    The API spans ISA-95 Levels 1-4:
    - **Level 1-2** - Process control, sensors, equipment monitoring
    - **Level 3** - MES core: work orders, scheduling, quality, traceability, performance, labor
    - **Level 4** - ERP integration: S&OP, purchase orders, order-to-cash, ISA-95 messaging
    - **Cross-Level** - AI/ML intelligence layer (proposed MESA F12)

    All responses use JSON. Paginated endpoints return an envelope with
    `{data, page, per_page, total}`.
  version: 4.0.0
  contact:
    name: Stephen Eacuello
    email: seacuello@uri.edu
  license:
    name: Academic Use
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/workorders/list:
    get:
      tags: ["Level 3 - F1 Resource Allocation"]
      summary: List work orders (paginated)
      description: |
        Returns paginated work orders with product details, ordered by due date ascending.
        Supports standard pagination via `page` and `per_page` query parameters.
      operationId: listWorkOrders
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/per_page"
      responses:
        "200":
          description: Paginated work order list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/WorkOrder"
  /api/workorders/create:
    post:
      tags: ["Level 3 - F1 Resource Allocation"]
      summary: Create a new work order
      description: |
        Creates a work order with auto-generated WO ID and lot number.
        Validates that the product exists and quantity is positive.
        Logs creation to the audit trail.
      operationId: createWorkOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_id]
              properties:
                product_id:
                  type: string
                  description: Must reference an existing product
                wo_id:
                  type: string
                  description: Optional; auto-generated if omitted (WO-2026-NNN)
                order_qty_kft:
                  type: number
                  description: Order quantity in thousands of feet
                due_date:
                  type: string
                  format: date
                priority:
                  type: integer
                  default: 3
                  minimum: 1
                  maximum: 5
                business_unit:
                  type: string
                  default: Industrial
            example:
              product_id: "PROD-A-001"
              order_qty_kft: 25.0
              due_date: "2026-05-15"
              priority: 2
              business_unit: "Industrial"
      responses:
        "200":
          description: Work order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  wo_id:
                    type: string
                  lot_number:
                    type: string
              example:
                ok: true
                wo_id: "WO-2026-146"
                lot_number: "LOT-2026-0301"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/workorders/{wo_id}:
    get:
      tags: ["Level 3 - F1 Resource Allocation"]
      summary: Work order detail
      description: |
        Returns comprehensive work order detail including operations, quality
        records (NCRs), test results, lot genealogy, scrap, holds, and audit trail.
      operationId: getWorkOrder
      parameters:
        - name: wo_id
          in: path
          required: true
          schema:
            type: string
          example: "WO-2026-001"
      responses:
        "200":
          description: Full work order detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  work_order:
                    $ref: "#/components/schemas/WorkOrder"
                  operations:
                    type: array
                    items:
                      type: object
                  quality:
                    type: array
                    items:
                      type: object
                  tests:
                    type: array
                    items:
                      type: object
                  lots:
                    type: array
                    items:
                      type: object
                  scrap:
                    type: array
                    items:
                      type: object
                  holds:
                    type: array
                    items:
                      type: object
                  audit:
                    type: array
                    items:
                      type: object
        "404":
          description: Work order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/workorders/update_status:
    post:
      tags: ["Level 3 - F1 Resource Allocation"]
      summary: Update work order status
      description: |
        Transitions a work order to a new status. Sets `actual_start` when
        moving to InProcess and `actual_end` when moving to Complete.
        Valid statuses: Pending, InProcess, Complete, On Hold, Cancelled.
      operationId: updateWorkOrderStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wo_id, status]
              properties:
                wo_id:
                  type: string
                status:
                  type: string
                  enum: [Pending, InProcess, Complete, On Hold, Cancelled]
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  wo_id:
                    type: string
                  status:
                    type: string

  # ====================================================================
  # Level 3 - F2 Scheduling
  # ====================================================================
...
components:
  parameters:
    page:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        default: 1
        minimum: 1
    per_page:
      name: per_page
      in: query
      description: Items per page (max 200)
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
  schemas:
    PaginatedEnvelope:
      type: object
      properties:
        data:
          type: array
          items: {}
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
      required: [data, page, per_page, total]
    WorkOrder:
      type: object
      properties:
        wo_id:
          type: string
        product_id:
          type: string
        order_qty_kft:
          type: number
        priority:
          type: integer
        due_date:
          type: string
          format: date
        planned_start:
          type: string
        planned_end:
          type: string
        actual_start:
          type: string
          nullable: true
        actual_end:
          type: string
          nullable: true
        status:
          type: string
          enum: [Pending, InProcess, Complete, On Hold, Cancelled]
        business_unit:
          type: string
        created_date:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string